% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agemoments.R
\name{agemoments}
\alias{agemoments}
\title{agemoments}
\usage{
agemoments(age)
}
\arguments{
\item{age}{A variable containing ages (e.g., radiocarbon dates)}
}
\value{
A vector containing the following components:
  \item{mean}{arithmetic mean of specimen ages}
  \item{median}{median of specimen ages}
  \item{sdev}{standard deviation of specimen ages}
  \item{iqr}{interquartile age range of specimen ages}
  \item{g1}{sample skewness estimate g1 of specimen ages}
  \item{t3}{sample L-skewness estimate t3 of specimen ages}
}
\description{
Computes univariate statistics for a sample of
individually dated specimens.
}
\details{
The function returns basic descriptive univariate statistics,
including mean, median, standard deviation, interquartile range,
sample skewness, and L-skewness
}
\examples{

agemoments(bahamas$spec$carbonateAge)

}
