\name{bahamas}
\alias{bahamas}
\docType{data}
\title{bahamas: 60 specimens dated using radiocarbon}
\description{An example of individually dated invertebrate specimens
collected from a single shallow water site in Bahamas}
\format{A list with two objects.
\describe{
\item{spec}{matrix with 60 rows (specimens) with raw c14 dates from single sites and two species}
\item{spec}{matrix with posterior-calibrated Baysian age probabilities for the same 60 specimens}
 }
}
\details{A dataset including 30 mollusk shells and 30 echinoid tests from 1 sampling station off the coast of San Salvador, Bahamas. Data from Kowalewski et al. (2018).}
\references{Kowalewski, M., Casebolt*, S., Hua, Q., Whitacre, K.E.,
Kaufman, D., Kosnik, M.A., (2018), One fossil record, multiple
time-resolutions: Comparative time-averaging of mollusks and echinoids on a carbonate platform. Geology 46: 51-54.}
\source{ \url{https://doi.org/10.1130/G39789.1}}
\keyword{datasets}
\examples{str(bahamas)}
