\name{florida}
\alias{florida}
\docType{data}
\title{florida dataset}
\description{An example dataset of individually dated invertebrate specimens
collected from multiple sites in Cedar Key, Florida}
\format{A list with two objects.
\describe{
\item{spec}{matrix with with raw c14 dates}
\item{post}{matrix with posterior-calibrated Baysian age probabilities for the same specimens}
 }
}
\details{A dataset including radiocarbon ages for mollusk shells, echinoid tests, mollusk fragments, and echinoid samples from Cedar Key, Florida}
\references{none so far}
\keyword{datasets}
\examples{str(florida)}
