% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{moments}
\usage{
moments(age)
}
\arguments{
\item{age}{A variable containing raw ages (e.g., radiocarbon ages)}
}
\value{
A vector containing the following components:
  \item{medianage}{median age of specimens in a given sample}
}
\description{
moments
}
