% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postdist.R
\name{postdist}
\alias{postdist}
\title{postdist}
\usage{
postdist(age, prob, id, size = 10000)
}
\arguments{
\item{age}{A numerical vector containing age values to which probabilities are assigned}

\item{prob}{A numerical vector containing probabilities for each age value}

\item{id}{A vector containing unique specimen ids}

\item{size}{a single numerical value (default = 10000)  defining how many age values
should be sampled to approximate the posterior probability distribution}
}
\value{
A single matrix of the class 'postdist':
  \item{out}{returns a numerical matrix with rows representing individual
  age dates columns representing Monte Carlo ages sampled from posterior
  probability distributions}
}
\description{
Generates specimen/date-level Monte Carlo distributions of age uncertainties
based on Bayesian posterior probabilities of ages
}
\details{
Radiocarbon calibration provides posterior probabilities of age values
given the observed uncalibrated radiocarbon ages. This function generates a
Monte Carlo approximation of posterior probability distribution converting probabilities
into a predefined number of age values representing age uncertainty around
a given date.
}
\examples{

out1 <- postdist(age=bahamas$post$Age, prob=bahamas$post$Probability,
id=bahamas$post$Specimen)

dim(out1)

}
