% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeaveraging.R
\docType{package}
\name{timeaveraging}
\alias{timeaveraging-package}
\alias{timeaveraging}
\title{timeaveraging: assessing age distributions of fossil specimens}
\description{
The package 'timeaveraging' provides functions to assess and visualize
age distributions while incorporating dating uncertainty estimated via
Bayesian posterior calibrations (specimen-level age probability
distributions)
}
\details{
The current version of the package provides functions for
assessing moments of age distributions
}
\examples{

# data examples
data(bahamas)
str(bahamas)

data(florida)
str(florida)

# postdist function
postdist(age=bahamas$post$Age, prob=bahamas$post$Probability,
id=bahamas$post$Specimen)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mjkowalewski/timeaveraging}
  \item Report bugs at \url{https://github.com/mjkowalewski/timeaveraging}
}

}
\author{
Michal Kowalewski \email{kowalewski@ufl.edu}
}
