% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeavestats.R
\name{timeavestats}
\alias{timeavestats}
\title{timeavestats}
\usage{
timeavestats(x, times = 100)
}
\arguments{
\item{x}{A list of class 'postdist' returned by 'postdist' function}

\item{times}{A numerical value indicating number of replicate
bootstrap samples desired}
}
\value{
A dataframe with sample-level statistics
}
\description{
Evaluates confidence intervals around time-averaging statistics
using bootstrapping while incorporating dating uncertainty.
Accepts object of the class 'postidst' generated by function 'postdist'.
}
\details{
The function timeavestats returns sample-level observed
and bootstrap statistics including naive bootstrap 95% confidence intervals

The following parameters are returned:

observed sample stats: mean (mean age), median (median age),
TAV (total variance), TAVexact ("exact" total variance), AEV (age estimation
variance), AEVexact ("exact" age estimation variance), ETA (estimated time
averaging), ETAexact ("exact" estimated time averaging), OTA (uncorrected
estimate of time averaging - 2 x std.dev), iqr (interquartile range of ages),
g1 (parametric sample skeweness estimate), t3 (l-skewness)

corresponding boostrap statistics with 95% naive bootstrap confidence intervals.
}
\examples{

out1 <- postdist(age=florida$post$cal.BP, prob=florida$post$probability,
id=florida$post$UniqueID, idname='id', size=10000, outdata = TRUE)

timeavestats(out1, times=100)

}
\references{
Hosking, JRM, 2023, L-Moments. R package, version 3.0.
https://CRAN.R-project.org/package=lmom

Ritter, M, Erthal F, Kosnik, MA, Kowalewski, M, Coimbra, CJ,
Caron, F, Kaufman, DS, 2023, Onshore-offshore trends in the temporal
resolution of molluscan death assemblages: how age-frequency
distributions reveal Quaternary sea-level history.
Palaios, 38: 148-157. http://dx.doi.org/10.2110/palo.2021.041
}
