% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeaveraging.R
\docType{package}
\name{timeveraging}
\alias{timeaveraging-package}
\alias{timeveraging}
\alias{timeaveraging}
\title{timeaveraging: assessing age distirbutions of individually dated fossil specimens}
\description{
timeaveraging package provides common measures of age distributions
while taking in consideration dating uncertainty estimated via
Bayesian posterior calibrations (specimen-level age probability
distributions)
}
\details{
Plot functions to visualize age distributions are also provided.

The current version of the package provides functions for
assessing moments of age distributions
}
\examples{

# agemoments
agemoments(bahamas$spec$carbonateAge)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mjkowalewski/timeaveraging}
  \item Report bugs at \url{https://github.com/mjkowalewski/timeaveraging}
}

}
\author{
Michal Kowalewski \email{kowalewski@ufl.edu}
}
