% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postdist.R
\name{postdist}
\alias{postdist}
\title{postdist}
\usage{
postdist(age, prob, id, idname = NULL, size = 10000, outdata = FALSE)
}
\arguments{
\item{age}{A numerical vector containing age values to which probabilities are assigned}

\item{prob}{A numerical vector containing probabilities for each age value}

\item{id}{A vector containing unique specimen ids}

\item{idname}{A character string proving a label for 'id' variable (optional)}

\item{size}{a single numerical value (default = 10000)  defining how many age values
should be sampled to approximate the posterior probability distribution}

\item{outdata}{logical (default=FALSE) to indicate whether generated samples
should be returned. If FALSE, a single data.frame with statistical descriptors
is returned. If TRUE, a list of two date.frames is returned.}
}
\value{
A list with two data.frames (if outdate=TRUE) and a data.frame (if outdate=False):
  \item{stats}{returns means, variances and medians of posterior age distributions
   and exact means and medians calculated directly from the probability distribution}
  \item{outdata}{archives Monte Carlo approximations of posterior probability distributions,
  returned if outdata=TRUE}
}
\description{
Generates specimen/date-level Monte Carlo distributions of age uncertainties
based on Bayesian posterior probabilities of ages
}
\details{
Radiocarbon calibration provides posterior probabilities of age values
given the observed uncalibrated radiocarbon age. This function generates a Monte Carlo
approximation of posterior probability distribution converting probabilities into
a predefined number of age values representing age uncertainty around a given date.
The number of distributions equals the number of dated specimens. The function preserves
the order of id values in the original file.
}
\examples{

postdist(age=bahamas$post$Age, prob=bahamas$post$Probability,
id=bahamas$post$Specimen, idname='Specimen')

}
