% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tastats.R
\name{tastats}
\alias{tastats}
\title{tastats}
\usage{
tastats(x, use.median = T, simple = T)
}
\arguments{
\item{x}{A list of class 'postdist' returned by \code{\link{postdist}} function}

\item{use.median}{A logical argument (default=TRUE) to determine whether to
use median age (TRUE) or mean age (FALSE) to estimate central tendency of the
posterior distribution of age probabilities}

\item{simple}{A logical argument (default=TRUE) to determine if additional
information should be returned by the function.}
}
\value{
A dataframe with sample-level statistics (if simple=T) or
a list with four items (if simple=F):
  \item{my.stats}{returns a dataframe (if simple = T) or the first item in
  the list of items (if simple = F). This is a single set of numerical values
  providing summary statistics for a given sample}
  \item{ages}{returns age estimates (mean or median ages) for individual
  observations, if simple = F}
  \item{n}{returns the number of age dates in a sample, if simple = F}
  \item{num.replic.ages}{returns the number of mcmc replicate samples
  from the posterior probability distibution of each age date, if simple = F}
  \item{age.estimate}{returns the information if median or mean of the
  posterior probability distribution was used to derive an estimated
  specimen age, if simple = F}
}
\description{
Returns summary of statistics describing an age distribution of
age dates. Accepts object of the class 'postidst' generated by function
\code{\link{postdist}}
}
\details{
If argument 'simple' is set to TRUE (default), the function returns
a single-row data.frame with summary statistics for the analyzed sample.
If argument 'simple' is set to FALSE, the function returns a list

The following summary statistics are returned:

median.age - median age of a sample calculated from mean or median ages
of posterior distributions of individual ages

mean.age - mean age of a sample calculated from mean or median ages
of posterior distributions of individual ages

std.age - standard deviation of a sample calculated from mean or median ages
of posterior distributions of individual ages

iq.age - interquartile range calculated from mean or median ages of posterior
distributions of individual ages

AEV - Age Estimation Variance calculated as average of variances of posterior
Monte-Carlo distributions of individual ages included in a sample
(see Ritter et al. 2023)

TAV - Total Assemblage Variance calculated as variance of all Monte-Carlo
ages (see Ritter et al. 2023)

ETA (ES) - Estimated Time-Averaging (2 standard deviations) corrected for
dating uncertainty calculated as 2 * square root of TAV-AEV
(see Ritter et al. 2023)

g1 - skewness, the third moment about the mean calculated from mean or
median ages of posterior distributions

l3 - L-skeweness calculated using function \code{\link[lmom]{samlmu}}
from package "lmom".
}
\examples{

out1 <- postdist(age=florida$post$cal.BP, prob=florida$post$probability,
id=florida$post$UniqueID, size=10000)

tastats(out1)

}
\references{
Hosking, JRM, 2023, L-Moments. R package, version 3.0.
https://CRAN.R-project.org/package=lmom

Ritter, M, Erthal F, Kosnik, MA, Kowalewski, M, Coimbra, CJ,
Caron, F, Kaufman, DS, 2023, Onshore-offshore trends in the temporal
resolution of molluscan death assemblages: how age-frequency
distributions reveal Quaternary sea-level history.
Palaios, 38: 148-157. http://dx.doi.org/10.2110/palo.2021.041
}
