% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeavestats.R
\name{timeavestats}
\alias{timeavestats}
\title{timeavestats}
\usage{
timeavestats(x, times = 100)
}
\arguments{
\item{x}{A list of class 'postdist' returned by \code{\link{postdist}} function}

\item{times}{A numerical value indicating number of replicate
bootstrap samples desired}
}
\value{
A dataframe with sample-level statistics
}
\description{
Evaluates confidence intervals around time-averaging statistics
using bootstrapping while incorporating dating uncertainty.
Accepts object of the class 'postidst' generated by function
\code{\link{postdist}}
}
\details{
The function returns sample-level statistics, including observed
values and bootstrapped values with confidence intervals

The following parameters are returned (for details and references
see \code{\link{postdist}}:

observed sample stats: mean (mean age), median (median age),
TAV (total variance), TAVexact ("exact" total variance), AEV (age estimation
variance), AEVexact ("exact" age estimation variance), ETA (estimated time
averaging), ETAexact ("exact" estimated time averaging), OTA (uncorrected
estimate of time averaging - 2 times std.dev), iqr (interquartile range of ages),
g1 (parametric sample skeweness estimate), t3 (l-skewness)

The corresponding bootstrap statistics with 95 percent naive bootstrap
confidence intervals are also included.
}
\examples{

out1 <- postdist(age=florida$post$cal.BP, prob=florida$post$probability,
id=florida$post$UniqueID, idname='id', size=10000, outdata = TRUE)

timeavestats(out1, times=100)

}
